Implement fixtures in BitcoindV17RpcClientTest fixtures
